requirements.txt
selenium==4.15.0
sqlalchemy==2.0.23
pandas==2.1.4
webdriver-manager==4.0.1
requests>=2.28.0
beautifulsoup4>=4.11.0
psycopg2-binary>=2.9.9
openpyxl>=3.1.2